Re-implemented search_bar functionality
search_bar file used to be dependent on chrome-driver which will be harder to mock than requests.
Also, requests is better from performance perspective as there is no need for GUI as drivers.
In order to work properly, I've added a new constant GOOGLE_SEARCH_URL and renamed this file as well its test